docs: update README files and remove benchmark documentation
- Revised the Chinese and English README files to enhance clarity and accuracy of performance benchmarks.
- Updated system specifications and performance test descriptions to reflect current hardware and Java versions.
- Removed outdated benchmark documentation files to streamline project structure and focus on essential information.
