+++ src/test/java/org/indunet/fastproto/benchmark/chain/Sample.java
